requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
azure-storage-blob
pydantic
requests
librosa
openai
